Move streamToUint8Array back to node-web-streams-helper.ts
